backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
beautifulsoup4==4.14.3
boto3==1.42.59
botocore==1.42.59
certifi==2026.2.25
charset-normalizer==3.4.4
click==8.3.1
distro==1.9.0
dnspython==2.8.0
fastapi==0.135.0
groq>=0.4.1,<1.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jmespath==1.1.0
lxml==6.0.2
pydantic==2.12.5
pydantic_core==2.41.5
pymongo==4.16.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
requests==2.32.5
s3transfer==0.16.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
tenacity>=8.2.0
mangum>=0.17.0

duckduckgo-search>=6.0.0
tenacity>=8.0.0
python-jose[cryptography]==3.3.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-groq>=0.2.0
opensearch-py>=2.6.0
langgraph>=0.2.20
